fixed product-routes.js delete request

+++ routes/api/product-routes.js
@@ -9,10 +9,11 @@ router.get('/', async (req, res) => {
   // be sure to include its associated Category and Tag data
   try {
     const productData = await Product.findAll({
-      //THESE ARE THE INCLUDED FOREIGN KEYS
       include: [{model: Category}, {model: Tag}]
     });
+
     res.status(200).json(productData);
+
   } catch (err) {
     res.status(500).json(err);
   }
@@ -115,14 +116,19 @@ router.put('/:id', (req, res) => {
 router.delete('/:id', async (req, res) => {
   // delete one product by its `id` value
   try {
-    const productData = await Product.destroy({
-      // include: [{model: Category}, {model: Tag}],
-      where: {
-        id: req.params.id,
-      }, 
+
+    const productData = await Product.findByPk(req.params.id, {
+      include:  [{model: Category}, {model: Tag}]
     });
 
-    console.log(req.params.id);
+    if(productData.id == req.params.id) {
+      Product.destroy({
+        // include: [{model: Category}, {model: Tag}],
+        where: {
+          id: req.params.id,
+        }, 
+      });
+    }
 
     if (!productData) {
       res.status(404).json({ message: 'No product with this id!' });
